% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_contingency_table.R
\name{calculate_contingency_table}
\alias{calculate_contingency_table}
\title{Calculate contingency table}
\usage{
calculate_contingency_table(MBH_table)
}
\arguments{
\item{MBH_table}{dataframe of mutual best hits with genomic coordinates loaded by the load_MBH_table function}
}
\value{
A gdataframe object
}
\description{
This is a function to generate the contingency table of an MBH dataframe and apply fischer test to calculate the significant associations.
}
