% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_macrosynt.R
\name{calculate_macrosynt}
\alias{calculate_macrosynt}
\title{Calculate contingency table}
\usage{
calculate_macrosynt(mbh_df, pvalue_threshold = 0.001)
}
\arguments{
\item{mbh_df}{dataframe. mutual best hits with genomic coordinates loaded with load_mbh_df()}

\item{pvalue_threshold}{numeric. threshold for significancy. (default equals 0.001)}
}
\value{
A dataframe object
}
\description{
This is a function to generate the contingency table of an MBH dataframe and
apply fischer test to calculate the significant associations. It outputs a datafram shaped as following :
sp1.Chr,sp2.Chr,orthologs,pval
}
