% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mbh_df.R
\name{load_mbh_df}
\alias{load_mbh_df}
\title{load Mutual Best Hit table.}
\usage{
load_mbh_df(mbh_table, sp1_bed, sp2_bed)
}
\arguments{
\item{mbh_table}{string. Full path to the mutual best hits table (format : geneID_on_species1   geneID_on_species2)}

\item{sp1_bed}{string. Full path to the genomic coordinates of the genes on species1 (BED format)}

\item{sp2_bed}{string. Full path to the genomic coordinates of the genes on species2 (BED format)}
}
\value{
dataframe composed of genomic coordinates of orthologs on both species
}
\description{
Puts together the mutual best hits calculated beforehand with the genomic coordinates of the orthologs on both species to be compared.
}
