% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_MBH_table.R
\name{load_MBH_table}
\alias{load_MBH_table}
\title{load Mutual Best Hit table.}
\usage{
load_MBH_table(
  MBH_table_path,
  species1_bed_path,
  species2_bed_path,
  sp1_keep_n_chr = 0,
  sp2_keep_n_chr = 0,
  sp1_keep_chr_names = NULL,
  sp2_keep_chr_names = NULL
)
}
\arguments{
\item{MBH_table_path}{path to the mutual best hits table (format : geneID_on_species1   geneID_on_species2)}

\item{species1_bed_path}{path to the genomic coordinates of the genes on species1 (BED format)}

\item{species2_bed_path}{path to the genomic coordinates of the genes on species2 (BED format)}

\item{sp1_keep_n_chr}{integer to tell the n chromosomes of species1 with the higher amount of orthologs to keep (default is to keep all)}

\item{sp2_keep_n_chr}{integer to tell the n chromosomes of species2 with the higher amount of orthologs to keep (default is to keep all)}

\item{sp1_keep_chr_names}{list of chr names to keep on species1.(default is to keep all)}

\item{sp2_keep_chr_names}{list of chr names to keep on species 2. (default is to keep all)}
}
\value{
dataframe composed of genomic coordinates of orthologs on both species
}
\description{
Puts together the mutual best hits calculated beforehand with the genomic coordinates of the orthologs on both species to be compared.
}
