% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fischer_test.R
\name{plot_fischer_test}
\alias{plot_fischer_test}
\title{Plot fischer test}
\usage{
plot_fischer_test(
  contingency_table,
  sp1_name = "",
  sp2_name = "",
  plot_legend = FALSE
)
}
\arguments{
\item{contingency_table}{dataframe of contingency table with p-values calculated by the calculate_contingency_table function}

\item{sp1_name}{string. The name of the species1 to display on the plot}

\item{sp2_name}{string. The name of the species2 to put on the plot}

\item{plot_legend}{boolean set to TRUE to display the legend (default = FALSE)}
}
\value{
A ggplot2 object
}
\description{
This is a function to generate the contingency table of an MBH dataframe and apply fischer test to calculate the significant associations.
}
