% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_macrosynt.R
\name{plot_macrosynt}
\alias{plot_macrosynt}
\title{Plot Macro-synteny}
\usage{
plot_macrosynt(macrosynt_df, sp1_label = "", sp2_label = "")
}
\arguments{
\item{macrosynt_df}{dataframe of contingency table with p-values calculated by the calculate_contingency_table function}

\item{sp1_label}{string. The name of the species1 to display on the plot}

\item{sp2_label}{string. The name of the species2 to put on the plot}
}
\value{
ggplot2 object
}
\description{
This is a function to generate the contingency table of an MBH dataframe and apply fischer test to calculate the significant associations.
}
