% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oxford_grid.R
\name{plot_oxford_grid}
\alias{plot_oxford_grid}
\title{plot the Macro-synteny as an oxford grid.}
\usage{
plot_oxford_grid(
  mbh_df,
  sp1_label = "",
  sp2_label = "",
  dot_size = 0.2,
  colors = FALSE
)
}
\arguments{
\item{mbh_df}{dataframe. mutual best hits with genomic coordinates loaded by the load_mbh_df()}

\item{sp1_label}{string. name of 1st species to display on the plot}

\item{sp2_label}{string. name of 2nd species to display on the plot}

\item{dot_size}{float. default = 0.2}

\item{colors}{logical. default = FALSE}
}
\value{
A ggplot2 object
}
\description{
This is a function to plot the oxford grided plot to compare the macro synteny of two species. Its input will have been loaded using load_mbh_df()
}
