% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_synteny_oxford_grid.R
\name{plot_synteny_oxford_grid}
\alias{plot_synteny_oxford_grid}
\title{plot the synteny oxford grid.}
\usage{
plot_synteny_oxford_grid(
  MBH_table,
  colors = FALSE,
  sp1_name = "",
  sp2_name = "",
  sp1_keep_chr_names = NULL,
  sp2_keep_chr_names = NULL,
  sp1_chr_order = NULL,
  sp2_chr_order = NULL,
  dot_size = 0.2
)
}
\arguments{
\item{MBH_table}{dataframe of mutual best hits with genomic coordinates loaded by the load_MBH_table function}

\item{sp1_name}{string. name of 1st species to display on the plot}

\item{sp2_name}{string. name of 2nd species to display on the plot}

\item{sp1_keep_chr_names}{list of chr names to keep on species1.(default is to keep all)}

\item{sp2_keep_chr_names}{list of chr names to keep on species 2. (default is to keep all)}

\item{sp1_chr_order}{ordered list of chromosome names to use for plotting}

\item{sp2_chr_order}{ordered list of chromosome names to use for plotting}
}
\value{
A ggplot2 object
}
\description{
This is a function to plot the oxford grided plot to compare the macro synteny of two species. Its input will have been loaded using the load_MBH_table function, and the output is a ggplot2 object.
}
