% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_synteny.R
\name{reorder_synteny}
\alias{reorder_synteny}
\title{Reorder the mbh_df before plotting}
\usage{
reorder_synteny(orthologs_df, pvalue_threshold = 0.001)
}
\arguments{
\item{orthologs_df}{dataframe. mutual best hits with genomic coordinates loaded with load_orthologs()}

\item{pvalue_threshold}{numeric. threshold for significancy. (default equals 0.001)}
}
\value{
A dataframe object
}
\description{
his is a function to reorder an orthologs_df, that was generated with load_orthologs(). It allows for automatic reordering, or manual filtering and reordering.
}
\seealso{
\code{\link[=load_orthologs]{load_orthologs()}}

\code{\link[=compute_macrosynteny]{compute_macrosynteny()}}
}
